part 3 - organize code
