+++ jwt_auth/server.js
@@ -2,7 +2,8 @@ require('dotenv').config()
 
 const express = require('express')
 const app = express()
-const jwt = require('jsonwebtoken')
+
+const { authenticateToken } = require('./middlewares/authenticate')
 
 const PORT = 3000
 
@@ -26,45 +27,10 @@ app.get('/posts', authenticateToken, (req, res) => {
     res.json(posts.filter(post => post.username === req.user.name ))
 })
 
-app.post('/login', (req, res) => {
-    // auth user
+// AUTH logic moved to authServer
+// app.post('/login', (req, res) => {...}
 
-    const username = req.body.username
-    const user = { name: username }
-
-    // create JWT
-    // serialize user object with a secret
-    const accessToken = jwt.sign(user, process.env.ACCESS_TOKEN_SECRET)
-    res.json({ accessToken })
-
-})
-
-// authenticate middleware
-function authenticateToken(req, res, next) {
-    // verify that this is a user
-
-    // get token from header
-    // Authorization: Bearer TOKEN
-
-    const authHeader = req.headers['authorization']
-    const token = authHeader?.split(' ')[1]
-
-    if (!token) {
-        // no credentials
-        return res.sendStatus(401)
-    }
-
-    // verify token
-    jwt.verify(token, process.env.ACCESS_TOKEN_SECRET,
-        (err, user) => {
-            if (err) {
-                // invalid credentials
-                res.sendStatus(401)
-            }
-            req.user = user
-            next()
-        })
-
-}
+// MIDDLEWARE moved to middlewares
+// function authenticateToken(req, res, next) {...}
 
 app.listen(PORT)
